+++ src/headerSection/headerSection.js
@@ -1,24 +1,18 @@
-import buttonAddDeck from "./buttonAddDeck/buttonAddDeck.js";
-import buttonOpenDeck from "./buttonOpenDeck/buttonOpenDeck.js";
-import buttonSelectDecks from "./buttonSelectDecks/buttonSelectDecks.js";
-import buttonRemoveDecks from "./buttonRemoveDecks/buttonRemoveDecks.js";
-import deckNameQuerries from "./deckNameQuerries/deckNameQuerries.js";
+import addDeckSection from "./addDeckSection/addDeckSection";
+import deckSection from "./deckSection/decksSection";
+import deckSection from "./deckSection/decksSection";
+import selectDecksSection from "./selectDecksSection/selectDecksSection";
 
 export default function headerSection() {
   localStorage.clear();
   const header = document.querySelector("#header");
   const bigFunctionStack = new Object();
-  const deckSection = document.querySelector("#deckSection");
-  const addDeckSection = document.querySelector("#addDeck");
-  const selectDecksSection = document.querySelector("#selectDecks");
-  const buttonAdd = buttonAddDeck(addDeck);
-  const buttonSelect = buttonSelectDecks(selectDecks);
-  const buttonRemove = buttonRemoveDecks(removeDecks);
-  let deckList = new Array();
-  const db = deckNameQuerries();
-  let deckInFocus;
+  const handleDeckSection = deckSection(loadFlashCards, showFooter);
+  const handleAddDeckSection = addDeckSection(addDeck);
+  const handleSelectDecksSection = selectDecksSection(getSelectDecksCallBacks);
 
   function setCallBacks(functionStack) {
+    deckSectionCallBacks = {};
     bigFunctionStack.loadFlashCards = functionStack.loadFlashCards;
     bigFunctionStack.hideContent = functionStack.hideContent;
     bigFunctionStack.toggleAllExcept = functionStack.toggleAllExcept;
@@ -26,113 +20,43 @@ export default function headerSection() {
   }
 
   function addDeck() {
-    const loadFlashCards = bigFunctionStack.loadFlashCards;
-    const showFooter = bigFunctionStack.footer.show;
+    handleDeckSection.addDeck();
+  }
+
+  function loadFlashCards() {
+    bigFunctionStack.loadFlashCards();
+  }
+
+  function showFooter() {
+    bigFunctionStack.showFooter();
+  }
+
+  function getSelectDecksCallBacks() {
+    const toggleAllExcept = bigFunctionStack.toggleAllExcept;
+    const enableButtonAdd = handleAddDeckSection.enable;
+    const disableButonnAdd = handleAddDeckSection.disable;
+    const setDeleteMode = handleDeckSection.setDeleteMode;
+    const removeDecks = handleDeckSection.removeDecks;
+
     const functionStack = {
-      setDeckInFocus,
-      loadFlashCards,
-      showFooter,
+      toggleAllExcept,
+      enableButtonAdd,
+      disableButonnAdd,
+      setDeleteMode,
+      removeDecks,
+      resetDeckView,
     };
-    const defaultName = "NewDeck";
-    const deckName = generateName(defaultName);
-    const deckObject = {};
-    db.insertDeckObject(deckName, deckObject);
-    const buttonHandle = buttonOpenDeck(deckName, functionStack);
-    const button = buttonHandle.button;
-    deckSection.appendChild(button);
-    deckList.push(buttonHandle);
-  }
 
-  function generateName(name) {
-    const nameList = db.getAllDeckNames();
-    const defaultName = name;
-    for (let i = 1; nameList.includes(name); i++) {
-      name = defaultName + i;
-    }
-    return name;
-  }
-
-  function setDeckInFocus(deckName) {
-    deckInFocus = deckName;
-  }
-
-  function getDeckInFocus() {
-    return deckInFocus;
-  }
-
-  function selectDecks() {
-    const toggleAllExcept = bigFunctionStack.toggleAllExcept;
-    toggleAllExcept("header");
-    disableButonnAdd();
-    buttonAdd.disabled = true;
-    deckList.forEach((buttonHandle) => buttonHandle.setDeleteMode());
-    selectDecksSection.removeChild(buttonSelect);
-    selectDecksSection.appendChild(buttonRemove);
-  }
-
-  function removeDecks() {
-    const toggleAllExcept = bigFunctionStack.toggleAllExcept;
-    toggleAllExcept("header");
-    enableButtonAdd();
-    buttonAdd.disabled = false;
-    deckList.forEach((buttonHandle) => {
-      buttonHandle.setNormalMode();
-      const deleteFlag = buttonHandle.getDeleteFlag();
-      if (deleteFlag) removeDeck(buttonHandle);
-    });
-    selectDecksSection.removeChild(buttonRemove);
-    selectDecksSection.appendChild(buttonSelect);
-    resetDeckView();
-  }
-
-  function enableButtonAdd() {
-    addDeckSection.appendChild(buttonAdd);
-    addDeckSection.style.backgroundColor = "white";
-  }
-
-  function disableButonnAdd() {
-    buttonAdd.remove();
-    addDeckSection.style.backgroundColor = "black";
-  }
-
-  function removeDeck(buttonHandle) {
-    const button = buttonHandle.button;
-    const deckName = buttonHandle.getDeckName();
-    db.removeDeckObject(deckName);
-    button.remove();
-    deckList = deckList.filter(deck => deck.getDeckName() != deckName);
-    console.log(deckList);
+    return functionStack;
   }
 
   function resetDeckView() {
-    const nameList = db.getAllDeckNames();
-    if (!nameList.includes(deckInFocus)) {
+    if (handleDeckSection.currentDeckGone()) {
       bigFunctionStack.hideContent();
       bigFunctionStack.footer.hide();
     }
   }
-
-  function changeDeckName(deckNameOld, deckName) {
-    if (deckName !== deckNameOld) {
-      const buttonHandle = getButtonHandle(deckNameOld);
-      const deckNameNew = generateName(deckName);
-      db.updateDeckObject(deckNameOld, deckNameNew);
-      buttonHandle.setDeckName(deckNameNew);
-      if(deckNameOld == getDeckInFocus())
-        setDeckInFocus(deckNameNew);
-    }
-    return deckName;
-  }
-
-  function getButtonHandle(deckName) {
-    let myButtonHanndle;
-    deckList.forEach((buttonHandle) => {
-      if (buttonHandle.getDeckName() == deckName)
-        myButtonHanndle = buttonHandle;
-    });
-    return myButtonHanndle;
-  }
-
+ 
   function enableAll() {
     header.style.pointerEvents = "auto";
     header.style.backgroundColor = "white";
