+++ src/headerSection/selectDecksSection/buttonRemoveDecks/buttonRemoveDecks.js
